Only add "return" to print REPL results when needed
The "return" keyword is added in the buffer only when it is not already in
the statement introduced by the user, and when the statement is not an
assignment. Handling the first case allows errors when using the history
from readline, the second one is needed because assignments do not compute
a result.