Set chisel.options from command line arguments
Picks non-flag arguments and stores them in the chisel.options table. For
"key=value" options, the "value" is stored in the table under the given
"key", as a string. For other options, they are set in the table as items
with a "true" value associated to them.

This allows passing down command line options to Lua scripts which are
run by chisel.