Add printerdata:_init() method
The initilizer for "printerdata" will pick the options from the "options"
table, and convert each of them to proper "printeroption" objects. For each
option, it tries to call "_init_option_<optionname>", which should pick the
table describing the option and return an object. Then the object replaces
the table with the data.