Make printerdata.get() load pure tabular data
To avoid having code to run when loading device information,
printerdata.get() will now just load a table from the data files.
To keep the possibility of having inheritance in device data files,
if the "base" attribute is defined, it is assumed to contain the
name of the base device.