chisel-ppd: CUPS-compatible driver program
Implements a chisel-ppd tool that can be used to list supported devices and
generate PPD files. It is suitable to be used as a driver for cups-driverd.
The implementation is quite naïve, as it uses printerdata.get() to load each
one of the data files. If needed, a more clever approach (e.g. a cache)
might be added in the future, but for now the script gets the job done.