chisel-ppd: Allow listing the plain device IDs
Using "chisel-ppd list plain" will produce a listing in stdout containing
only the device identifiers. This is faster as it does not incur in opening
all device data files, and is useful for scripting, e.g. for looping over
the list of devices and creating all the PPD files at once.