Skip listing devices with names starting with underscore
chisel-ppd will skip listing devices whose manufacturer-id or model-id
start with an underscore. This is useful to provide base devices which
would not render a valid device description, and thus should not be
listed in the output of the "chisel-ppd list" comand.

Prior to this, skipping was done by checking both manufacturer and model
for non-nil values. The problem with that approach is that the base
(i.e. incomplete) devices would still appear in the output of
"chisel-ppd list plain".