loader: Validate document options
Contains an initial implementation of the document options validator.
A table contains the recognized options as keys, with the associated
values being the functions used to convert/validate the values. The
later is convenient when options are given all as strings, still
some should be converted to other types (mostly useful with numbers)