ibv4: Implement dot_distance option
This allso pulls in some support code needed to handle options: For each
option in document.options, an <option_name>_option method will be tried;
if the method exists, it is called passing the option value, if not, the
option is ignored.

The ibv4:dot_distance_option() method will search for the closest match
to the requested dot distance in the table of supported values. This
means that it is a best-effort implementation.