chisel-ppd: Use "chisel-ppd" as model prefix on listing
In device lists generated with "chisel-ppd list", the prefix used
in the name of the device must match the name of the program installed
under the "driver" directory of CUPS. The prefix is used as name of
the program used to generate the PPDs with the "cat" command. If the
prefix and the executable name do not match, CUPS will not be able
to obtain PPDs.