boot: Remove implicit call to _init on cloned objects
Calling the _init() method implicitly makes it so in some cases, depending
on how object:clone() is called the _init() method would still not be
called. To get rid of possible indefined behaviour, no methods will be
called implicitly, and if a constructor is needed it should be explicitly
invoked, e.g.:

  spam = object:clone {
    init = function (self)
      print ("being initialized...")
      return self
    end
  }

  eggs = spam:clone { name = "eggs" } :init ()

As a bonus of making "constructors" explicit, they may have any number of
arguments.