printerdata: Gather default values in a table
Default values for device options will be gathered in the "default"
attribute of "printerdata" instances, so for example to access the
default value for the "dot_distance" setting one would:

  data = lib.printerdata.get (devicename)
  print (data.default.dot_distance)

Defaults are stored in a separate table to ease iterating over the
values for the default settings. This information is meant to be
used at device initialization.