device: Allow defining options with no PPD equivalent
Device data options which are just plain values i.e. they do not extend the
device.option class, or they do not provide a :ppd() method. Those options
are just not translated in the generated PPD files. This is needed to be
able to specify defaults for options in the device data files.