device: Implement device:create_renderer()
Provided that devices know which renderer has to be used for sending data to
them, it is convenient that they have a method for creating them. As a bonus
using the new method will set the "device" attribute to the device object
used to instantiate the renderer:

  dev  = lib.device.get ("...")
  rend = dev:create_renderer ()
  assert (dev == rend.device)

This will make possible for renderers to know the exact device they are
outputting data to, which may be needed to properly implement renderers.